Split twitchApi.ts: extract OAuth and EventSub subscription functions into twitchApiEventSub.ts
